requirements.txt
pyaes==1.6.1
pyasn1==0.4.8
python-decouple==3.5
rsa==4.7.2
Telethon==1.23.0
